src/main/java/pageObjects/android/PlacementPrep.java
package pageObjects.android;

import org.openqa.selenium.WebElement;
import org.openqa.selenium.support.PageFactory;

import io.appium.java_client.android.AndroidDriver;
import io.appium.java_client.pagefactory.AndroidFindBy;
import io.appium.java_client.pagefactory.AppiumFieldDecorator;

public class PlacementPrep {
	
AndroidDriver driver;
	
	public PlacementPrep(AndroidDriver driver) {
		this.driver = driver;
		PageFactory.initElements(new AppiumFieldDecorator(driver), this);
	}
	
	
	@AndroidFindBy(xpath="//android.view.ViewGroup[@resource-id=\"RouteScreen: 0\"]/android.view.ViewGroup/android.view.ViewGroup/android.view.ViewGroup/android.view.ViewGroup/android.view.ViewGroup/android.view.ViewGroup/android.view.ViewGroup/android.view.ViewGroup[2]/android.widget.ScrollView/android.view.ViewGroup/android.view.ViewGroup[2]/android.widget.ScrollView/android.view.ViewGroup/android.view.ViewGroup[1]")
	WebElement placementPrepCard;
	
	@AndroidFindBy(xpath="//android.widget.TextView[@text=\"Go To Content\"]")
	WebElement goToContentButton;
	
	
	
	
	public void openCard() {
		placementPrepCard.click();
	}
	
	public void goToContent() {
		goToContentButton.click();
	}
}
